+++ spec/lib/glimmer/data_binding/model_binding_spec.rb
@@ -109,6 +109,44 @@ describe Glimmer::DataBinding::ModelBinding do
       expect(person.siblings).to eq(new_siblings)
     end
     
+    it 'reads and writes changes in a hash attribute' do
+      model_binding = described_class.new(person, :hash_attribute)
+      
+      Glimmer::DataBinding::Observer.proc do |new_value|
+        @observer_notified = true
+      end.observe(model_binding)
+      
+      person.hash_attribute[:d] = 4
+        
+      expect(@observer_notified).to be_truthy
+      
+      new_value = {e: 5}
+      model_binding.call(new_value)
+        
+      expect(person.hash_attribute).to eq(new_value)
+    end
+    
+    it 'reads and writes changes in a hash attribute key' do
+      model_binding = described_class.new(person.hash_attribute, :d)
+      
+      Glimmer::DataBinding::Observer.proc do |new_value|
+        @observer_notified = true
+      end.observe(model_binding)
+      
+      person.hash_attribute[:c] = 4
+        
+      expect(@observer_notified).to be_falsey
+      
+      person.hash_attribute[:d] = 4
+        
+      expect(@observer_notified).to be_truthy
+      
+      new_value = 5
+      model_binding.call(new_value)
+        
+      expect(person.hash_attribute[:d]).to eq(new_value)
+    end
+    
     it 'reads and writes changes in an indexed model' do
       person.siblings = []
       model_binding = described_class.new(person, 'siblings[0]')